% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_token.R
\name{get_microsoft365_token}
\alias{get_microsoft365_token}
\title{Retrieve Microsoft 365 Access Token}
\usage{
get_microsoft365_token(
  scope,
  tenant = read_secret("azure.tenant"),
  app_id = read_secret("azure.app_id"),
  auth_type = read_secret("azure.auth_type"),
  ...,
  error_on_fail = FALSE
)
}
\arguments{
\item{scope}{this must be "outlook", "teams", "planner", "tasks" (which is "planner" without group rights), or "sharepoint", and will set the right API permission for each}

\item{tenant}{the tenant to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{app_id}{the Azure app id to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{auth_type}{the authentication method to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{...}{other arguments passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}
}
\description{
This function uses retrieves an access token from the department's Microsoft 365 account.
}
