% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{git_compare}
\alias{git_compare}
\title{Git Compare Files}
\usage{
git_compare(files = NULL, original_file = NULL)
}
\arguments{
\item{files}{character vector, allowed to be named}

\item{original_file}{original file path to restore file to}
}
\description{
This compares two text files
}
