% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{project_consult_add}
\alias{project_consult_add}
\title{Add Project Or Consult Using Shiny}
\usage{
project_consult_add(
  planner = connect_planner(),
  teams = if (in_positron()) connect_teams() else NULL,
  channel = if (in_positron()) teams_projects_channel() else NULL,
  outlook = connect_outlook()
)
}
\arguments{
\item{planner}{Microsoft Planner account, as returned by e.g. \code{\link[=connect_planner]{connect_planner()}}}

\item{teams}{Microsoft Teams account, as returned by e.g. \code{\link[=connect_teams]{connect_teams()}}}

\item{channel}{Microsoft Teams Channel folder, as returned by e.g. \code{\link[=teams_projects_channel]{teams_projects_channel()}}}

\item{outlook}{MIcrosoft Outlook account, as returned by e.g. \code{\link[=connect_outlook]{connect_outlook()}}}
}
\description{
This is a Shiny app to add a new project: it creates a project folder locally \link[=connect_teams]{or in Teams}, generates the required Quarto or R Markdown or R file, and creates a \link[=connect_planner]{new task in Planner}. These functions come with RStudio addins to quickly access existing projects. For consults, it only adds a Planner task and creates the card in the background.
}
