% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_shiny.R
\name{project_add}
\alias{project_add}
\alias{project_edit}
\title{Add Project Using Shiny}
\usage{
project_add(
  board = read_secret("trello.default.board"),
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token"),
  planner = planner_connect(),
  teams = teams_connect(),
  projects_channel = read_secret("teams.projects.channel")
)

project_edit(
  card_number = project_get_current_id(ask = TRUE),
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)
}
\arguments{
\item{board}{Trello board}

\item{username}{Trello username}

\item{key}{Trello API key}

\item{token}{Trello API token}

\item{planner}{Microsoft Planner account, as returned by e.g. \code{\link[=planner_connect]{planner_connect()}}}

\item{teams}{Microsoft Teams account, as returned by e.g. \code{\link[=teams_connect]{teams_connect()}}}

\item{projects_channel}{name of a channel in the Teams account set in \code{teams}}

\item{card_number}{Trello card number}
}
\description{
This is a Shiny app to add a new project: it creates a project folder, generates the required Quarto or R Markdown or R files, and creates a new card in Trello. These functions come with RStudio addins to quickly access existing projects.
}
