% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pin_list}
\alias{pin_meta}
\alias{pin_versions}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pins}{\code{\link[pins:pin_list]{pin_list()}}, \code{\link[pins:pin_meta]{pin_meta()}}, \code{\link[pins:pin_versions]{pin_versions()}}}
}}

