% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{sharepoint}
\alias{sharepoint}
\alias{download_from_sharepoint}
\alias{upload_to_sharepoint}
\alias{source_sharepoint}
\alias{download_file_to_sharepoint}
\title{Upload to / Download from SharePoint Project Channel}
\usage{
download_from_sharepoint(
  remote_file_name = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  account = connect_teams(),
  overwrite = TRUE
)

upload_to_sharepoint(
  local_file_name = NULL,
  remote_file_name = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  account = connect_teams()
)

source_sharepoint(
  remote_file_name = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  account = connect_teams(),
  ...
)

download_file_to_sharepoint(
  url,
  local_file_name = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  account = connect_teams(),
  ...
)
}
\arguments{
\item{remote_file_name}{Name of the file to search, will be used together with \code{project_number} and can be a regular expression. Will be passed on to \code{\link[=project_get_file]{project_get_file()}}.}

\item{project_number}{Planner project number}

\item{full_sharepoint_path}{Full path of the remote file, consisting of the project folder and the file name, such as \code{"Project - p123/my_file.R"}.}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_teams]{connect_teams()}} or \code{\link[Microsoft365R:get_team]{Microsoft365R::get_team()}}.}

\item{overwrite}{Whether to overwrite an existing file.}

\item{local_file_name}{File name of the local file.}

\item{...}{arguments passed on to \code{\link[=source]{source()}} or \code{\link[=download.file]{download.file()}}}

\item{url}{URL to download and afterwards upload to SharePoint}
}
\description{
These functions allow for fast and easy transfer between local files and SharePoint, using \code{\link[=teams_projects_channel]{teams_projects_channel()}} to retrieve the Projects channel.
}
\details{
You must either provide \code{remote_file_name} + \code{project_number}, or provide \code{full_sharepoint_path}.
\itemize{
\item \code{download_from_sharepoint()} finds a SharePoint file in the \link[=teams_projects_channel]{Projects channel} using \code{\link[=project_get_file]{project_get_file()}}, and downloads it to a local temporary folder. The local filename is returned and can be used in other functions.
\item \code{upload_to_sharepoint()} uploads any local file to SharePoint in the \link[=teams_projects_channel]{Projects channel}. The remote path is returned and can be used in other functions.
\item \code{source_sharepoint()} runs \code{\link[=source]{source()}} on a SharePoint project file, by downloading it to a temporary folder first using \code{download_from_sharepoint()}.
\item \code{download_file_to_sharepoint()} downloads a remote URL and then uploads the downloaded file to SharePoint using \code{upload_to_sharepoint()}
}
}
\examples{
\dontrun{

source_sharepoint("my_file.R", 123)

"my_file.R" |>
  download_from_sharepoint(project_number = 123) |>
  source_sharepoint()
  
url <- "https://examplefile.com/file-download/26"
from_internet <- download_file_to_sharepoint(url,
                                             "test-download.pdf",
                                             project_number = 123)
from_internet
#> [1] "Testproject - p123/test-download.pdf"

local_file <- download_from_sharepoint(full_sharepoint_path = from_internet)
}
}
