% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certeprojects-package}
\alias{certeprojects}
\alias{certeprojects-package}
\title{certeprojects: A Certe R Package for Department Projects}
\description{
A Certe R Package with functions to automate department projects using Shiny, MS Teams and Trello. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certeprojects}
  \item \url{https://github.com/certe-medical-epidemiology/certeprojects}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
  \item Erwin Dijkstra \email{e.dijkstra@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
