% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_connect.R
\name{connect}
\alias{connect}
\alias{get_microsoft365_token}
\alias{connect_outlook}
\alias{connect_planner}
\alias{connect_teams}
\title{Connect to Microsoft 365}
\usage{
get_microsoft365_token(
  scope = read_secret("azure.scope_list"),
  tenant = read_secret("azure.tenant"),
  app_id = read_secret("azure.app_id"),
  auth_type = read_secret("azure.auth_type"),
  ...,
  overwrite = TRUE,
  error_on_fail = TRUE
)

connect_outlook(email = read_secret("mail.auto_from"), overwrite = FALSE, ...)

connect_planner(
  plan = read_secret("planner.name"),
  team_name = read_secret("team.name"),
  overwrite = FALSE,
  ...
)

connect_teams(team_name = read_secret("team.name"), overwrite = FALSE, ...)
}
\arguments{
\item{scope}{any (combination) of "outlook", "teams", "planner", "tasks" (which is "planner" without group rights), or "sharepoint". Can also be a manual vector of Microsoft Permissions, such as \verb{"User.Read}.}

\item{tenant}{the tenant to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{app_id}{the Azure app id to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{auth_type}{the authentication method to use, passed on to \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}}

\item{...}{arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}

\item{overwrite}{a \link{logical} to overwrite an existing connection, useful for switching accounts}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}

\item{email}{email address of the user, or a shared mailbox}

\item{plan}{name of the team's plan if \code{team_name} is not empty. Otherwise, a plan ID (for individual use).}

\item{team_name}{name of the team, can be left blank to connect to an individual planner}
}
\description{
These functions create a connection to Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{planner_*()} and \verb{teams_*()} functions allow to work with these connections.
}
\details{
\subsection{Microsoft Outlook}{

To switch between different Outlook accounts, run \code{\link[=connect_outlook]{connect_outlook()}} with another email address, and set \code{overwrite = TRUE}. This will allow all \code{certemail} functions to use the newly set account.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# at default connects to the department mailbox:
connect_outlook()
# afterwards, this does nothing since `overwrite` is not set:
connect_outlook("user@certe.nl")
# this switches to the user account:
connect_outlook("user@certe.nl", overwrite = TRUE)
}\if{html}{\out{</div>}}

Using \code{overwrite} is needed, because running just \code{\link[=connect_outlook]{connect_outlook()}} afterwards again (which many \code{certemail} functions do) will otherwise change back to the default account.
}

\subsection{Microsoft Planner}{

To connect to MS Planner with a personal account, retrieve the plan ID (e.g., from the URL of the plan), and pass it on to \code{\link[=connect_planner]{connect_planner()}} as \code{plan}, and set \code{overwrite = TRUE} over replace an existing connection. Make sure that \code{team_name} is left blank:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{connect_planner(plan = "AAA-0aa0AaAa-aaaAAAAAAAa", team_name = NULL, overwrite = TRUE)
}\if{html}{\out{</div>}}
}

\subsection{Microsoft Teams}{

Connecting to MS Teams can only be done on the group (= team) level. It is not possible to set up a connection without a valid team name.
}

\subsection{Pre-loaded Settings}{

When attaching this \code{certeprojects} package using \code{\link[=library]{library()}}, and external R process will be run in the background using the \code{callr} package to connect to MS Outlook, MS Planner, and MS Teams. This will increase speed when connecting using \code{\link[=connect_outlook]{connect_outlook()}}, \code{\link[=connect_planner]{connect_planner()}}, or \code{\link[=connect_teams]{connect_teams()}}.
}
}
