% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_microsoft365_token.R
\name{get_microsoft365_token}
\alias{get_microsoft365_token}
\title{Retrieve Microsoft 365 Access Token}
\usage{
get_microsoft365_token(
  scope,
  shared_mbox_email = read_secret("mail.auto_from"),
  tenant = read_secret("mail.tenant"),
  app_id = read_secret("mail.app_id"),
  auth_type = read_secret("mail.auth_type"),
  error_on_fail = FALSE
)
}
\arguments{
\item{scope}{this must be "mail", "teams", or "planner", and will set the right API permission for each}

\item{shared_mbox_email}{email address of the shared mailbox to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{tenant}{the tenant to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{app_id}{the Azure app id to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{auth_type}{the authentication method to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}
}
\description{
This function uses retrieves an access token from the department's Microsoft 365 account.
}
