% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{git_compare}
\alias{git_compare}
\title{Git Compare Files}
\usage{
git_compare(files = NULL, file_title = NULL)
}
\arguments{
\item{files}{character vector, allowed to be named}

\item{file_title}{character to use as file title}
}
\description{
This compares two text files
}
