% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{pins}
\alias{pins}
\alias{export_pin}
\alias{import_pin}
\alias{remove_pin}
\alias{board_certeprojects}
\title{Work with Pins}
\usage{
export_pin(
  x,
  name = NULL,
  title = NULL,
  type = NULL,
  description = NULL,
  board = board_certeprojects()
)

import_pin(name, version = NULL, hash = NULL, board = board_certeprojects())

remove_pin(name, version = NULL, board = board_certeprojects())

board_certeprojects(
  site_name = read_secret("teams.name"),
  delete_by_item = TRUE,
  ...
)
}
\arguments{
\item{x}{An object (typically a data frame) to pin.}

\item{name}{Pin name.}

\item{title}{A title for the pin; most important for shared boards so that
others can understand what the pin contains. If omitted, a brief
description of the contents will be automatically generated.}

\item{type}{File type used to save \code{x} to disk. Must be one of
"csv", "json", "rds", "arrow", or "qs". If not supplied, will use JSON for
bare lists and RDS for everything else. Be aware that CSV and JSON are
plain text formats, while RDS, Arrow, and
\href{https://CRAN.R-project.org/package=qs}{qs} are binary formats.}

\item{description}{A detailed description of the pin contents.}

\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_connect]{board_connect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{version}{Retrieve a specific version of a pin. Use \code{\link[pins:pin_versions]{pin_versions()}} to
find out which versions are available and when they were created.}

\item{hash}{Specify a hash to verify that you get exactly the dataset that
you expect. You can find the hash of an existing pin by looking for
\code{pin_hash} in \code{\link[pins:pin_meta]{pin_meta()}}.}

\item{site_name}{Name of the SharePoint site}

\item{delete_by_item}{Whether to handle folder deletions on an item-by-item
basis, rather than deleting the entire folder at once. You may need to set
this to \code{TRUE} for a board in SharePoint Online or OneDrive for Business,
due to document protection policies that prohibit deleting non-empty
folders.}

\item{...}{Arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}
}
\description{
These functions can be used to work with \href{https://pins.rstudio.com}{pins}, developed by RStudio.
}
\details{
The board will be automatically retrieved based on the Microsoft 365 account (by running \code{\link[=board_certeprojects]{board_certeprojects()}}), leading to the SharePoints folder "pins" within the Microsoft 365 account.

For Pins functions of the \code{pins} package, use \code{\link[=board_certeprojects]{board_certeprojects()}} as board, e.g.:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{  pin_list(board_certeprojects())
}\if{html}{\out{</div>}}

The following \code{pins} functions are re-exported by this package: \code{\link[=pin_list]{pin_list()}}, \code{\link[=pin_meta]{pin_meta()}}, and \code{\link[=pin_versions]{pin_versions()}}.
}
