% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_planner.R
\name{planner}
\alias{planner}
\alias{planner_connect}
\alias{planner_get_bucket}
\alias{planner_create_bucket}
\alias{planner_get_task}
\alias{planner_create_task}
\title{Connect to Microsoft Planner via Microsoft 365}
\usage{
planner_connect(
  team_name = read_secret("teams.name"),
  plan_name = read_secret("planner.name"),
  ...
)

planner_get_bucket(
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect()
)

planner_create_bucket(bucket_name, account = planner_connect())

planner_get_task(task_title, task_id = NULL, account = planner_connect())

planner_create_task(
  title,
  descr = "",
  due = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect()
)
}
\arguments{
\item{team_name}{name of the team}

\item{...}{arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}

\item{bucket_name}{name of the bucket}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=planner_connect]{planner_connect()}} or via \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}.}

\item{task_title}{title of the task}

\item{task_id}{ID of the task}

\item{title, descr, due, assigned}{properties of the new task}
}
\description{
These functions create a connection to Microsoft Planner via Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{planner_*()} functions allow to work with this connection.
}
\examples{
\dontrun{

planner_upload("myfile.docx", channel = "My Channel")
planner_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
planner_upload(mtcars, channel = "My Channel")
mtcars |> 
  planner_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
planner_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
planner_download("myfile.docx", channel = "My Channel", card_number = 123)
planner_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- planner_import("mychannel/myfile.docx")
x <- planner_import("mychannel/myfile.docx", card_number = 123)
x <- planner_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
planner_open("test.xlsx", "My Channel")
planner_open("my channel/test.xlsx") # shorter version, tries to find channel
}
}
