% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_planner.R
\name{planner}
\alias{planner}
\alias{planner_connect}
\alias{planner_bucket_create}
\alias{planner_task_create}
\alias{planner_task_update}
\alias{planner_task_request_validation}
\alias{planner_task_validate}
\alias{planner_user_id}
\title{Connect to Microsoft Planner via Microsoft 365}
\usage{
planner_connect(
  team_name = read_secret("teams.name"),
  plan_name = read_secret("planner.name"),
  ...
)

planner_bucket_create(bucket_name, account = planner_connect())

planner_task_create(
  title,
  descr = NULL,
  duedate = NULL,
  requested_by = NULL,
  member = NULL,
  prio = read_secret("planner.default.priority"),
  checklist_items = NULL,
  checklist_name = read_secret("planner.default.checklist"),
  comments = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect()
)

planner_task_update(
  old_title,
  title = NULL,
  bucket_name = NULL,
  descr = NULL,
  duedate = NULL,
  assigned = NULL,
  assigned_keep = FALSE,
  categories = NULL,
  categories_keep = FALSE,
  comments = NULL,
  account = planner_connect()
)

planner_task_request_validation(
  title,
  category_text = read_secret("planner.label.authorise"),
  account = planner_connect()
)

planner_task_validate(
  title,
  category_text = read_secret("planner.label.authorised"),
  account = planner_connect()
)

planner_user_id(
  user,
  team_name = read_secret("teams.name"),
  account = planner_connect()
)
}
\arguments{
\item{team_name}{name of the team}

\item{...}{arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=planner_connect]{planner_connect()}} or via \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}.}

\item{title, descr, due, assigned}{properties of the new task}

\item{old_title}{title of the task to update}

\item{category_text}{text of the category to use}

\item{user}{a user name, mail adress, or Certe login name}
}
\description{
These functions create a connection to Microsoft Planner via Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{planner_*()} functions allow to work with this connection.
}
\examples{
\dontrun{

planner_upload("myfile.docx", channel = "My Channel")
planner_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
planner_upload(mtcars, channel = "My Channel")
mtcars |> 
  planner_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
planner_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
planner_download("myfile.docx", channel = "My Channel", card_number = 123)
planner_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- planner_import("mychannel/myfile.docx")
x <- planner_import("mychannel/myfile.docx", card_number = 123)
x <- planner_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
planner_open("test.xlsx", "My Channel")
planner_open("my channel/test.xlsx") # shorter version, tries to find channel
}
}
