% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_planner.R
\name{planner}
\alias{planner}
\alias{planner_connect}
\alias{planner_browse}
\alias{planner_bucket_create}
\alias{planner_buckets_list}
\alias{planner_tasks_list}
\alias{planner_categories_list}
\alias{planner_task_create}
\alias{planner_task_update}
\alias{planner_task_request_validation}
\alias{planner_task_validate}
\alias{planner_task_find}
\alias{planner_user_property}
\title{Connect to Microsoft Planner via Microsoft 365}
\usage{
planner_connect(
  plan = read_secret("planner.name"),
  team_name = read_secret("team.name"),
  ...
)

planner_browse(account = planner_connect())

planner_bucket_create(bucket_name, account = planner_connect())

planner_buckets_list(account = planner_connect(), plain = FALSE)

planner_tasks_list(account = planner_connect(), plain = FALSE)

planner_categories_list(account = planner_connect())

planner_task_create(
  title,
  description = NULL,
  duedate = NULL,
  requested_by = NULL,
  priority = read_secret("planner.default.priority"),
  checklist_items = NULL,
  categories = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect(),
  add_project_number = FALSE
)

planner_task_update(
  task,
  title = NULL,
  bucket_name = NULL,
  description = NULL,
  duedate = NULL,
  assigned = NULL,
  assigned_keep = FALSE,
  categories = NULL,
  categories_keep = FALSE,
  percent_completed = NULL,
  checklist_items = NULL,
  priority = NULL,
  attachment_urls = NULL,
  account = planner_connect()
)

planner_task_request_validation(
  task,
  category_text = read_secret("planner.label.authorise"),
  account = planner_connect()
)

planner_task_validate(
  task,
  category_text = read_secret("planner.label.authorised"),
  account = planner_connect()
)

planner_task_find(
  task_title = NULL,
  task_id = NULL,
  account = planner_connect()
)

planner_user_property(
  user,
  team_name = read_secret("team.name"),
  account = planner_connect(),
  property = "id",
  as_list = FALSE
)
}
\arguments{
\item{plan}{name of the team's plan if \code{team_name} is not empty. Otherwise, a plan ID (for individual use).}

\item{team_name}{name of the team}

\item{...}{arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=planner_connect]{planner_connect()}} or via \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}\verb{$get_group(name)$get_plan(plan_title)}.}

\item{bucket_name}{name of the bucket}

\item{plain}{return as plain names, not as \code{Azure} objects}

\item{title}{title of the task}

\item{description}{a description for the task. A vector will be add as one text separated by white lines.}

\item{duedate}{a date to use a due date, use \code{FALSE} to remove it}

\item{requested_by}{name of the person(s) who requested the task}

\item{priority}{a priority to set. Can be ranged between 0 (highest) and 10 (lowest), or: \code{"urgent"} or \code{"dringend"} for 1, \code{"important"} or \code{"belangrijk"} for 3, \code{"medium"} or \code{"gemiddeld"} or \code{FALSE} for 5, \code{"low"} or \code{"laag"} for 9. Priorities cannot be removed - the default setting is 5.}

\item{checklist_items}{character vector of checklist items}

\item{categories}{names of categories to add}

\item{assigned}{names of members within the plan. Use \code{FALSE} to remove all assigned members.}

\item{add_project_number}{a \link{logical} to add a project number}

\item{task}{any task title, task ID, or \code{\link[Microsoft365R:ms_plan_task]{ms_plan_task}} object (e.g. from \code{\link[=planner_task_find]{planner_task_find()}})}

\item{assigned_keep}{add members that are set in \code{assigned} instead of replacing them, defaults to \code{FALSE}}

\item{categories_keep}{add categories that are set in \code{categories} instead of replacing them, defaults to \code{FALSE}}

\item{percent_completed}{percentage of task completion between 0-100}

\item{attachment_urls}{URLs to add as attachment, can be named characters to give the URLs a title}

\item{category_text}{text of the category to use}

\item{task_title}{title of the task, will be searched with \code{\link[certetoolbox:like]{\%like\%}}}

\item{task_id}{exact id of the task}

\item{user}{a user name, mail adress, or Certe login name}

\item{property}{property to return, can be "id", "name" or "mail"}

\item{as_list}{return the full list of members as \link{list}, split into Eigenaars (Owners) / Leden (Members). This ignores \code{user}.}
}
\description{
These functions create a connection to Microsoft Planner via Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{planner_*()} functions allow to work with this connection.
}
\details{
The \code{\link[=planner_task_find]{planner_task_find()}} returns a \code{\link[Microsoft365R:ms_plan_task]{ms_plan_task}} object.
}
