% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_planner.R
\name{planner}
\alias{planner}
\alias{planner_connect}
\alias{planner_bucket_create}
\alias{planner_task_create}
\alias{planner_task_update}
\alias{planner_task_request_validation}
\alias{planner_task_validate}
\alias{planner_user_id}
\title{Connect to Microsoft Planner via Microsoft 365}
\usage{
planner_connect(
  team_name = read_secret("team.name"),
  plan_name = read_secret("planner.name"),
  ...
)

planner_bucket_create(bucket_name, account = planner_connect())

planner_task_create(
  title,
  descr = NULL,
  duedate = NULL,
  requested_by = NULL,
  priority = read_secret("planner.default.priority"),
  checklist_items = NULL,
  categories = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  account = planner_connect()
)

planner_task_update(
  old_title,
  title = NULL,
  bucket_name = NULL,
  descr = NULL,
  duedate = NULL,
  assigned = NULL,
  assigned_keep = FALSE,
  categories = NULL,
  categories_keep = FALSE,
  checklist_items = NULL,
  priority = NULL,
  account = planner_connect()
)

planner_task_request_validation(
  title,
  category_text = read_secret("planner.label.authorise"),
  account = planner_connect()
)

planner_task_validate(
  title,
  category_text = read_secret("planner.label.authorised"),
  account = planner_connect()
)

planner_user_id(
  user,
  team_name = read_secret("team.name"),
  account = planner_connect()
)
}
\arguments{
\item{team_name}{name of the team}

\item{plan_name}{name of the plan of the team}

\item{...}{arguments passed on to \code{\link[=get_microsoft365_token]{get_microsoft365_token()}}}

\item{bucket_name}{name of the bucket}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=planner_connect]{planner_connect()}} or via \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}.}

\item{title}{title of the task}

\item{descr}{a description for the task}

\item{duedate}{a date to use a due date, use \code{FALSE} to remove it}

\item{requested_by}{name of the person who requested the task}

\item{priority}{a priority to set. Can be ranged between 0 (highest) and 10 (lowest), or: \code{"urgent"} or \code{"dringend"} for 1, \code{"important"} or \code{"belangrijk"} for 3, \code{"medium"} or \code{"gemiddeld"} or \code{FALSE} for 5, \code{"low"} or \code{"laag"} for 9. Priorities cannot be removed - the default setting is 5.}

\item{checklist_items}{character vector of checklist items}

\item{categories}{names of categories to add}

\item{assigned}{names of members within the plan. Use \code{FALSE} to remove all assigned members.}

\item{old_title}{title of the task to update}

\item{assigned_keep}{add members that are set in \code{assigned} instead of replacing them, defaults to \code{FALSE}}

\item{categories_keep}{add categories that are set in \code{categories} instead of replacing them, defaults to \code{FALSE}}

\item{category_text}{text of the category to use}

\item{user}{a user name, mail adress, or Certe login name}
}
\description{
These functions create a connection to Microsoft Planner via Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{planner_*()} functions allow to work with this connection.
}
\examples{
\dontrun{

planner_upload("myfile.docx", channel = "My Channel")
planner_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
planner_upload(mtcars, channel = "My Channel")
mtcars |> 
  planner_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
planner_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
planner_download("myfile.docx", channel = "My Channel", card_number = 123)
planner_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- planner_import("mychannel/myfile.docx")
x <- planner_import("mychannel/myfile.docx", card_number = 123)
x <- planner_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
planner_open("test.xlsx", "My Channel")
planner_open("my channel/test.xlsx") # shorter version, tries to find channel
}
}
