% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_planner.R
\name{planner}
\alias{planner}
\alias{planner_browse}
\alias{planner_bucket_create}
\alias{planner_buckets_list}
\alias{planner_task_create}
\alias{planner_task_update}
\alias{planner_tasks_list}
\alias{planner_task_search}
\alias{planner_task_find}
\alias{planner_categories_list}
\alias{planner_retrieve_project_id}
\alias{planner_task_request_validation}
\alias{planner_task_validate}
\alias{planner_user_property}
\alias{planner_highest_project_id}
\alias{as.data.frame.ms_object}
\alias{as_tibble.ms_object}
\title{Connect to Microsoft Planner via Microsoft 365}
\usage{
planner_browse(account = connect_planner())

planner_bucket_create(bucket_name, account = connect_planner())

planner_buckets_list(account = connect_planner(), plain = FALSE)

planner_task_create(
  title,
  description = NULL,
  startdate = Sys.Date(),
  duedate = NULL,
  requested_by = NULL,
  priority = read_secret("planner.default.priority"),
  checklist_items = NULL,
  categories = NULL,
  assigned = NULL,
  bucket_name = read_secret("planner.default.bucket"),
  attachment_urls = NULL,
  account = connect_planner(),
  project_number = planner_highest_project_id() + 1
)

planner_task_update(
  task,
  title = NULL,
  description = NULL,
  startdate = NULL,
  duedate = NULL,
  priority = NULL,
  checklist_items = NULL,
  categories = NULL,
  categories_keep = FALSE,
  assigned = NULL,
  assigned_keep = FALSE,
  bucket_name = NULL,
  percent_completed = NULL,
  attachment_urls = NULL,
  account = connect_planner()
)

planner_tasks_list(account = connect_planner(), plain = FALSE)

planner_task_search(
  search_term = ".*",
  limit = Inf,
  account = connect_planner()
)

planner_task_find(
  task_title = NULL,
  task_id = NULL,
  account = connect_planner()
)

planner_categories_list(account = connect_planner())

planner_retrieve_project_id(task, account = connect_planner())

planner_task_request_validation(
  task,
  category_text = read_secret("planner.label.authorise"),
  account = connect_planner()
)

planner_task_validate(
  task,
  category_text = read_secret("planner.label.authorised"),
  account = connect_planner()
)

planner_user_property(
  user,
  team_name = read_secret("team.name"),
  account = connect_planner(),
  property = "id",
  as_list = FALSE
)

planner_highest_project_id(
  task = read_secret("planner.dummycard"),
  account = connect_planner()
)

\method{as.data.frame}{ms_object}(
  x,
  row.names = NULL,
  optional = FALSE,
  account = connect_planner(),
  ...
)

\method{as_tibble}{ms_object}(x, account = connect_planner(), ...)
}
\arguments{
\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_planner]{connect_planner()}} or via \code{\link[AzureGraph:graph_login]{AzureGraph::create_graph_login()}}\verb{$get_group(name)$get_plan(plan_title)}.}

\item{bucket_name}{name of the bucket}

\item{plain}{return as plain names, not as \code{Azure} objects}

\item{title}{title of the task}

\item{description}{a description for the task. A vector of length > 1 will be added as one text separated by white lines.}

\item{startdate}{a date to use as start date, use \code{FALSE} to remove it. Defaults to today.}

\item{duedate}{a date to use as due date, use \code{FALSE} to remove it}

\item{requested_by}{name of the person(s) who requested the task, this will be added as first line to \code{description}}

\item{priority}{a priority to set. Can be ranged between 0 (highest) and 10 (lowest), or: \code{"urgent"} or \code{"dringend"} for 1, \code{"important"} or \code{"belangrijk"} for 3, \code{"medium"} or \code{"gemiddeld"} or \code{FALSE} for 5, \code{"low"} or \code{"laag"} for 9. Priorities cannot be removed - the default setting is 5.}

\item{checklist_items}{character vector of checklist items}

\item{categories}{names of categories to add, can be multiple, but must exactly match existing category names}

\item{assigned}{names of members within the plan - use \code{NULL} to not add members in \code{\link[=planner_task_create]{planner_task_create()}}, and use \code{FALSE} to remove all existing members in \code{\link[=planner_task_update]{planner_task_update()}}}

\item{attachment_urls}{URLs to add as attachment, can be named characters to give the URLs a title. If they are Excel, PowerPoint or Word files, a preview will be shown on the task card.}

\item{project_number}{the new project number to assign. Use \code{NULL} or \code{FALSE} to not assign a project number. Defaults to the currently highest project ID + 1.}

\item{task}{any task title, task ID, or \code{\link[Microsoft365R:ms_plan_task]{ms_plan_task}} object (e.g. from \code{\link[=planner_task_find]{planner_task_find()}})}

\item{categories_keep}{add categories that are set in \code{categories} instead of replacing them, defaults to \code{FALSE}}

\item{assigned_keep}{add members that are set in \code{assigned} instead of replacing them, defaults to \code{FALSE}}

\item{percent_completed}{percentage of task completion between 0-100}

\item{search_term}{search term, can contain regular expressions}

\item{limit}{maximum number of tasks to show}

\item{task_title}{title of the task, will be searched with \code{\link[certetoolbox:like]{\%like\%}}}

\item{task_id}{exact id of the task}

\item{category_text}{text of the category to use}

\item{user}{a user name, mail adress, or Certe login name}

\item{team_name}{name of the team, can be left blank to connect to an individual planner}

\item{property}{property to return, can be "id", "name" or "mail"}

\item{as_list}{return the full list of members as \link{list}, split into Eigenaars (Owners) / Leden (Members). This ignores \code{user}.}

\item{x}{an \code{ms_object}}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
These functions use the connection to Microsoft Planner set up with \code{\link[=connect_planner]{connect_planner()}}.
}
\details{
\code{\link[=planner_task_search]{planner_task_search()}} searches the title and description using case-insensitive regular expressions and returns an \code{\link[Microsoft365R:ms_plan_task]{ms_plan_task}} object. In interactive mode and with multiple hits, a menu will be shown to pick from.

\code{\link[=planner_task_find]{planner_task_find()}} searches task title or ID, and returns an \code{\link[Microsoft365R:ms_plan_task]{ms_plan_task}} object. It is used internally b a lot of \verb{planner_*} functions, very fast, and does not support interactive use.

\code{\link[=planner_retrieve_project_id]{planner_retrieve_project_id()}} retrieves the p-number from the task title and returns it as \link{integer}.

\code{\link[=planner_highest_project_id]{planner_highest_project_id()}} retrieves the currently highest project ID from the dummy card.

Using \code{\link[=as.data.frame]{as.data.frame()}} or \code{\link[=as_tibble]{as_tibble()}} on an \code{ms_object}, such as \code{ms_plan_task}, will return the properties and details of the object as a \link{data.frame}. For transforming many \code{ms_object}s to a data.frame, use \code{\link[=as.data.frame]{as.data.frame()}} or \code{\link[=as_tibble]{as_tibble()}} in \code{\link[=lapply]{lapply()}} and bind the list of objects together. For example, this retrieves a tibble with the properties and details of all tasks:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(dplyr)
planner_tasks_list() |> 
  lapply(as_tibble) |> 
  bind_rows()

# also works for other 'ms_object's, such as 'ms_channel':
teams_channels_list(plain = FALSE) |> 
  lapply(as_tibble) |> 
  bind_rows()
}\if{html}{\out{</div>}}
}
