% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{project_add}
\alias{project_add}
\alias{project_update}
\title{Add Project Using Shiny}
\usage{
project_add(
  planner = connect_planner(),
  teams = connect_teams(),
  channel = teams_projects_channel(account = teams)
)

project_update(
  current_task_id = project_get_current_id(ask = TRUE),
  planner = connect_planner()
)
}
\arguments{
\item{planner}{Microsoft Planner account, as returned by e.g. \code{\link[=connect_planner]{connect_planner()}}}

\item{teams}{Microsoft Teams account, as returned by e.g. \code{\link[=connect_teams]{connect_teams()}}}

\item{channel}{Microsoft Teams Channel folder, as returned by e.g. \code{\link[=teams_projects_channel]{teams_projects_channel()}}}

\item{current_task_id}{Project (p-)number of the project to update}
}
\description{
This is a Shiny app to add a new project: it creates a project folder locally \link[=connect_teams]{or in Teams}, generates the required Quarto or R Markdown or R file, and creates a \link[=connect_planner]{new task in Planner}. These functions come with RStudio addins to quickly access existing projects.
}
