% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_properties.R
\name{project_properties}
\alias{project_properties}
\alias{project_get_current_id}
\alias{project_identifier}
\alias{project_get_folder}
\alias{project_get_folder_full}
\alias{project_get_title}
\alias{project_get_file}
\alias{project_set_file}
\alias{project_set_folder}
\alias{project_open_analysis_file}
\alias{project_open_folder}
\title{Project Properties}
\usage{
project_get_current_id(ask = NULL)

project_identifier(card_number = project_get_current_id())

project_get_folder(card_number = project_get_current_id())

project_get_folder_full(card_number = project_get_current_id())

project_get_title(card_number = project_get_current_id())

project_get_file(
  filename,
  card_number = project_get_current_id(),
  fixed = FALSE
)

project_set_file(filename, card_number = project_get_current_id())

project_set_folder(foldername, card_number = project_get_current_id())

project_open_analysis_file(card_number = project_get_current_id(ask = TRUE))

project_open_folder(card_number = project_get_current_id(ask = TRUE))
}
\arguments{
\item{ask}{logical to indicate whether the project number should always be asked. The default, \code{NULL}, will show a popup in \link[=interactive]{interactive} \R sessions, allowing to search for projects. In non-interactive sessions, such as in Quarto and R Markdown, it will use the current \link[=getwd]{working directory} to determine the project number.}

\item{card_number}{Planner card number}

\item{filename}{filename to set or get, case-insensitive, and can also be a \link[base:regex]{regular expression}}

\item{fixed}{\link{logical} to turn off regular expressions}

\item{foldername}{foldername to set}
}
\description{
Retrieve project properties, such as the title, folder location and project number.
}
\details{
\code{\link[=project_get_current_id]{project_get_current_id()}} uses \code{\link[=planner_task_search]{planner_task_search()}} to find a specific project based on any search string.

\code{\link[=project_identifier]{project_identifier()}} generates the project identifier for print on reports and in mails: a combination of the currently logged in user (in your case: '\Sexpr{Sys.info()["user"]}'), the current date/time (format: YYMMDDHHMM), and the project number. If the project number is not available, it will only return the current user and date/time (format: YYMMDDHHMM).

\code{\link[=project_set_folder]{project_set_folder()}} will create the folder if it does not exist.
}
\examples{
project_identifier(123)
}
