% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_properties.R, R/source.R
\name{project_properties}
\alias{project_properties}
\alias{project_get_current_id}
\alias{project_identifier}
\alias{project_get_folder}
\alias{project_get_title}
\alias{project_get_file}
\alias{project_set_file}
\alias{project_set_folder}
\alias{project_open_analysis_file}
\alias{project_open_folder}
\alias{project_add_qmd_skeleton}
\alias{get_output_folder}
\alias{get_output_folder_consult}
\alias{sharepoint_to_local_temp}
\alias{source_sharepoint}
\title{Project Properties}
\usage{
project_get_current_id(ask = NULL, account = connect_teams())

project_identifier(project_number = project_get_current_id())

project_get_folder(
  project_number = project_get_current_id(),
  account = connect_teams()
)

project_get_title(
  project_number = project_get_current_id(),
  account = connect_teams()
)

project_get_file(
  filename = ".*",
  project_number = project_get_current_id(),
  fixed = FALSE,
  account = connect_teams()
)

project_set_file(
  filename,
  project_number = project_get_current_id(),
  account = connect_teams()
)

project_set_folder(
  foldername,
  project_number = project_get_current_id(),
  account = connect_teams()
)

project_open_analysis_file(
  project_number = project_get_current_id(ask = TRUE),
  account = connect_teams()
)

project_open_folder(
  project_number = project_get_current_id(ask = TRUE),
  account = connect_teams()
)

project_add_qmd_skeleton(
  filename = NULL,
  project_number = project_get_current_id(),
  account = connect_planner()
)

get_output_folder(
  project_number = project_get_current_id(),
  account = connect_teams()
)

get_output_folder_consult()

sharepoint_to_local_temp(
  file_name = NULL,
  project_number = project_get_current_id(),
  full_path = NULL,
  account = connect_teams(),
  ...
)

source_sharepoint(full_path, account = connect_teams(), ...)
}
\arguments{
\item{ask}{logical to indicate whether the project number should always be asked. The default, \code{NULL}, will show a popup in \link[=interactive]{interactive} \R sessions, allowing to search for projects. In non-interactive sessions, such as in Quarto and R Markdown, it will use the current \link[=getwd]{working directory} to determine the project number.}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_teams]{connect_teams()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_team()}}.}

\item{project_number}{Planner project number}

\item{filename}{name for the new Quarto file}

\item{fixed}{\link{logical} to turn off regular expressions}

\item{foldername}{foldername to set}

\item{file_name}{Name of the file to search, will be used with \code{project_number}.}

\item{full_path}{Full path of file, including the project folder: \code{"Project - p123/my_file.R"}. Will be passed on to \code{\link[=sharepoint_to_local_temp]{sharepoint_to_local_temp()}}.}

\item{...}{arguments passed on to \code{\link[=source]{source()}}}
}
\description{
Retrieve project properties, such as the title, folder location and project number.

This runs \code{\link[=source]{source()}} on a SharePoint file, by downloading it to a temporary folder first.
}
\details{
\code{\link[=project_get_current_id]{project_get_current_id()}} uses \code{\link[=planner_task_search]{planner_task_search()}} to find a specific project based on any search string.

\code{\link[=project_identifier]{project_identifier()}} generates the project identifier for print on reports and in mails: a combination of the currently logged in user (in your case: '\Sexpr{Sys.info()["user"]}'), the current date/time (format: YYMMDDHHMM), and the project number. If the project number is not available, it will only return the current user and date/time (format: YYMMDDHHMM).

\code{\link[=project_set_folder]{project_set_folder()}} will create the folder if it does not exist.

\code{\link[=project_add_qmd_skeleton]{project_add_qmd_skeleton()}} initializes a new Quarto skeleton for a project.

\code{\link[=sharepoint_to_local_temp]{sharepoint_to_local_temp()}} finds a SharePoint file and downloads it to a local temporary folder. That local filename is returned and can be used in other functions, such as \code{\link[=source]{source()}}.
}
\examples{
project_identifier(123)
}
