% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_properties.R
\name{project_properties}
\alias{project_properties}
\alias{project_get_current_id}
\alias{project_identifier}
\alias{project_get_folder}
\alias{project_get_folder_full}
\alias{project_get_title}
\alias{project_get_file}
\alias{project_set_file}
\alias{project_set_folder}
\alias{project_open_analysis_file}
\alias{project_open_folder}
\title{Project Properties}
\usage{
project_get_current_id(ask = NULL)

project_identifier(card_number = project_get_current_id())

project_get_folder(card_number = project_get_current_id())

project_get_folder_full(card_number = project_get_current_id())

project_get_title(card_number = project_get_current_id())

project_get_file(
  filename,
  card_number = project_get_current_id(),
  fixed = FALSE
)

project_set_file(filename, card_number = project_get_current_id())

project_set_folder(foldername, card_number = project_get_current_id())

project_open_analysis_file(card_number = project_get_current_id(ask = TRUE))

project_open_folder(card_number = project_get_current_id(ask = TRUE))
}
\arguments{
\item{ask}{logical to indicate whether the project number should always be asked}

\item{card_number}{Trello card number}

\item{filename}{filename to set or get}

\item{fixed}{logical to turn off regular expressions}

\item{foldername}{foldername to set}
}
\description{
Retrieve project properties, such as the title, folder location and Trello card number.
}
\details{
\code{\link[=project_get_current_id]{project_get_current_id()}} uses \code{\link[=trello_search_card]{trello_search_card()}} to find a specific project based on any search string.

\code{\link[=project_identifier]{project_identifier()}} generates the project identifier for print on reports and in mails: a combination of the project number, the project creation date/time (format: YYMMDDHHMM) and the current date/time (format: YYMMDDHHMM). If the project number is not available, it will only return the current date/time (format: YYMMDDHHMM).

\code{\link[=project_set_folder]{project_set_folder()}} will create the folder if it does not exist.
}
