% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_properties.R
\name{project_properties}
\alias{project_properties}
\alias{project_get_current_id}
\alias{project_get_folder}
\alias{project_get_folder_full}
\alias{project_get_title}
\alias{project_get_file}
\alias{project_set_file}
\alias{project_set_folder}
\alias{project_open_analysis_file}
\title{Project Properties}
\usage{
project_get_current_id(ask = NULL)

project_get_folder(card_number = project_get_current_id())

project_get_folder_full(card_number = project_get_current_id())

project_get_title(card_number = project_get_current_id())

project_get_file(
  filename,
  card_number = project_get_current_id(),
  fixed = FALSE
)

project_set_file(filename, card_number = project_get_current_id())

project_set_folder(foldername, card_number = project_get_current_id())

project_open_analysis_file(card_number = project_get_current_id(ask = TRUE))
}
\arguments{
\item{ask}{logical to indicate whether the project number should always be asked}

\item{card_number}{Trello card number}

\item{filename}{filename to set or get}

\item{foldername}{foldername to set}

\item{fixed}{logical to turn off regular expressions}
}
\description{
Retrieve project properties, such as the title, folder location and Trello card number.
}
\details{
\code{\link[=project_set_folder]{project_set_folder()}} will create the folder if it does not exist.
}
