% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_add.R
\name{project_add}
\alias{project_add}
\title{Add Project Using Shiny}
\usage{
project_add(
  planner = planner_connect(),
  teams = NULL,
  projects_channel = read_secret("teams.projects.channel")
)
}
\arguments{
\item{planner}{Microsoft Planner account, as returned by e.g. \code{\link[=planner_connect]{planner_connect()}}}

\item{teams}{Microsoft Teams account, as returned by e.g. \code{\link[=teams_connect]{teams_connect()}}}

\item{projects_channel}{name of a channel in the Teams account set in \code{teams}}
}
\description{
This is a Shiny app to add a new project: it creates a project folder local \link[=teams_connect]{or in Teams}, generates the required Quarto or R Markdown or R files, and creates a \link[=planner_connect]{new task in Planner}. These functions come with RStudio addins to quickly access existing projects.
}
