% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\alias{knit}
\alias{render_sharepoint}
\title{Knit R Markdown or Quarto Document}
\usage{
render(input_file, output_file = NULL, quiet = FALSE, ...)

knit(input_file, output_file = NULL, quiet = FALSE, ...)

render_sharepoint(
  input_file = NULL,
  output_file = NULL,
  project_number = project_get_current_id(),
  full_sharepoint_path = NULL,
  quiet = FALSE,
  account = connect_teams(),
  output_folder = get_output_folder(project_number = project_number, account = account),
  upload = FALSE,
  ...
)
}
\arguments{
\item{input_file}{file to be rendered, can be R Markdown (\code{.Rmd}) or Quarto (\code{.qmd}), or a lot of other formats such as \code{.md}, \code{.ipynb} and \href{https://quarto.org/docs/reference/}{many other formats that Quarto supports}.}

\item{output_file}{The name of the output file. If using \code{NULL}, the output
filename will be based on the filename for the input file. \code{output_file} is
mapped to the \code{--output} option flag of the \code{quarto} CLI. It is expected to
be a filename only, not a path, relative or absolute.}

\item{quiet}{Suppress warning and other messages, from R and also Quarto CLI
(i.e \code{--quiet} is passed as command line).

\code{quarto.quiet} \R option or \code{R_QUARTO_QUIET} environment variable can be used to globally override a function call
(This can be useful to debug tool that calls \verb{quarto_*} functions directly).

On Github Actions, it will always be \code{quiet = FALSE}.}

\item{...}{arguments passed on to \code{\link[=render]{render()}}}

\item{project_number}{Planner project number}

\item{full_sharepoint_path}{Full path of the remote file, consisting of the project folder and the file name, such as \code{"Project - p123/my_file.R"}.}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_teams]{connect_teams()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_team()}}.}

\item{output_folder}{any output folder of the resulting output file, defaults to \code{\link[=get_output_folder]{get_output_folder()}}. Can also be \code{"sharepoint"}, which sets \code{upload = TRUE}.}

\item{upload}{whether to upload the resulting output file back to SharePoint, default is \code{FALSE}.}
}
\value{
As opposed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}, which always returns \code{NULL}, this \code{render()} function always returns the output file name (invisibly).
}
\description{
Runs input files through Quarto or RMarkdown, and allows Quarto to work with full paths (where Quarto itself requires relative paths). This function essentially replaces both \code{\link[rmarkdown:render]{rmarkdown::render()}} and \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.
}
\details{
Functions \code{\link[=knit]{knit()}} and \code{\link[=render]{render()}} are identical.

Any Quarto file (\code{.qmd}) will be rendered using \code{\link[quarto:quarto_render]{quarto::quarto_render()}}. Any RMarkdown file (\code{.Rmd}) will be rendered using \code{\link[rmarkdown:render]{rmarkdown::render()}}.

\code{\link[=render_sharepoint]{render_sharepoint()}} downloads a SharePoint project file to a local temporary folder, then renders it. Either \code{input_file} + \code{project_number} or \code{full_sharepoint_path} must be given.
}
