% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\alias{knit}
\title{Knit R Markdown or Quarto Document using Quarto}
\usage{
render(input_file, output_file = NULL, quiet = TRUE, as_job = "auto", ...)

knit(input_file, output_file = NULL, quiet = TRUE, as_job = "auto", ...)
}
\arguments{
\item{input_file}{file to be rendered, can be R Markdown (\code{.Rmd}) or Quarto (\code{.qmd}), or a lot of other formats such as \code{.md}, \code{.ipynb} and \href{https://quarto.org/docs/reference/}{many other formats that Quarto supports}.}

\item{output_file}{The name of the output file. If using \code{NULL}, the output
filename will be based on the filename for the input file. \code{output_file} is
mapped to the \code{--output} option flag of the \code{quarto} CLI. It is expected to
be a filename only, not a path, relative or absolute.}

\item{quiet}{Suppress warning and other messages.}

\item{as_job}{Render as an RStudio background job. Default is \code{"auto"},
which will render individual documents normally and projects as
background jobs. Use the \code{quarto.render_as_job} \R option to control
the default globally.}

\item{...}{other arguments passed on to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}
}
\description{
Runs input files through Quarto, and allows Quarto to work with full paths (where Quarto itself requires relative paths). This function essentially replaces both \code{\link[rmarkdown:render]{rmarkdown::render()}} and \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.
}
\details{
Functions \code{\link[=knit]{knit()}} and \code{\link[=render]{render()}} are identical.
}
