% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R, R/teams_legacy.R
\name{teams}
\alias{teams}
\alias{teams_connect}
\alias{teams_name}
\alias{teams_channels}
\alias{teams_view_sharepoint}
\alias{teams_send_message}
\alias{teams_upload}
\alias{teams_download}
\alias{teams_import}
\alias{teams_open}
\alias{teams_get_link}
\alias{teams_webhooks}
\title{Connect to Microsoft Teams via Microsoft 365}
\usage{
teams_connect(team_name = read_secret("teams.name"), ...)

teams_name(account = teams_connect())

teams_channels(account = teams_connect())

teams_view_sharepoint(channel, account = teams_connect())

teams_send_message(
  body,
  channel,
  content_type = c("text", "html"),
  attachments = NULL,
  account = teams_connect()
)

teams_upload(
  local_path,
  teams_path = basename(local_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_download(
  teams_path,
  local_path = basename(teams_path),
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_import(
  teams_path,
  card_number = project_get_current_id(ask = FALSE),
  channel = NULL,
  account = teams_connect()
)

teams_open(teams_path, channel = NULL, account = teams_connect())

teams_get_link(
  teams_path,
  share_type = c("view", "edit"),
  expire_after = "1 month",
  password = NULL,
  channel = NULL,
  account = teams_connect()
)

teams(
  message = "",
  channel,
  title = NULL,
  subtitle = NULL,
  items = NULL,
  items_title = NULL,
  images = NULL,
  images_title = NULL,
  buttons = NULL,
  message_thumbnail = NULL,
  markdown = TRUE,
  webhooks_file = read_secret("teams.webhooks.file"),
  ...
)

teams_webhooks(webhooks_file = read_secret("teams.webhooks.file"))
}
\arguments{
\item{team_name}{name of the team}

\item{...}{not used at the moment}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=teams_connect]{teams_connect()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_team()}}.}

\item{channel}{channel to send message to, run \code{\link[=teams_webhooks]{teams_webhooks()}} for list of valid channels}

\item{body}{text of the message}

\item{content_type}{type of content, must be "text" or "html"}

\item{attachments}{vector of file locations of attachments to add to the message}

\item{local_path}{file location on the local system, can also be a \link{data.frame} which will then be saved locally to the temp folder first}

\item{teams_path}{file location in Microsoft Teams, may also contain the channel name if \code{channel} is \code{NULL}, e.g., \code{teams_path = "channel name/test.xlsx"}}

\item{card_number}{Trello card number}

\item{share_type}{type of share, must be \code{"view"} (default) or \code{"edit"}}

\item{expire_after}{time span after which the share link expires, defaults to \code{"1 month"}, can also be e.g. \code{"7 days"}}

\item{password}{password to set for share link, defaults to blank}

\item{message}{message to send, supports markdown. Can also be a \link{data.frame}, which will be transformed using \code{\link[certestyle:plain_html_table]{certestyle::plain_html_table()}}.}

\item{title}{title for message}

\item{subtitle}{subtitle for message}

\item{items}{text items for in message, a named vector}

\item{items_title}{title to show above items}

\item{images}{URLs of images to add to card. Local images will be uploaded to Imgur.com, or will will be transformed using \code{\link[base64enc:base64]{base64enc::base64encode()}} if that fails.}

\item{images_title}{title to show above images}

\item{buttons}{buttons to show on card. Like \code{items}, this should be a named vector where the values are URLs (starting with \verb{http://} or \verb{https://}) or mail links (starting with \verb{mailto:})}

\item{message_thumbnail}{image to show on card}

\item{markdown}{turn markdown support on/off}

\item{webhooks_file}{the file locations of a YAML file with MS Teams webhooks in the form \code{channel: "URL"}}
}
\description{
These functions create a connection to Microsoft Teams via Microsoft 365 and saves the connection to the \code{certeprojects} package environment. The \verb{teams_*()} functions allow to work with this connection.

The older \code{\link[=teams]{teams()}} function relies on the \code{httr} package and requires \href{https://docs.microsoft.com/en-us/microsoftteams/platform/webhooks-and-connectors/how-to/add-incoming-webhook#add-an-incoming-webhook-to-a-teams-channel}{Incoming Webhooks} to be set in Teams.
}
\examples{
\dontrun{

teams_upload("myfile.docx", channel = "My Channel")
teams_upload("myfile.docx", "my channel/my folder/test.docx")

# also supports data frames, they will be saved locally in a temp folder
teams_upload(mtcars, channel = "My Channel")
mtcars |> 
  teams_upload(channel = "My Channel")
  
  
# integrates with department projects:

# upload "myfile.docx" from the project folder of p123 to "My Channel":
teams_upload("myfile.docx", channel = "My Channel", card_number = 123)
# download "myfile.docx" from "My Channel" to the project folder of p123:
teams_download("myfile.docx", channel = "My Channel", card_number = 123)
teams_download("mychannel/myfile.docx", card_number = 123)

# direct import from Teams requires the 'certetoolbox' package
x <- teams_import("mychannel/myfile.docx")
x <- teams_import("mychannel/myfile.docx", card_number = 123)
x <- teams_import("myfile.docx", channel = "My Channel", card_number = 123)

# open a file in Excel Online
teams_open("test.xlsx", "My Channel")
teams_open("my channel/test.xlsx") # shorter version, tries to find channel
}
\dontrun{

# send message using Webhooks
teams("test message", channel = "My Channel")
# send message using Microsoft 365 login
teams_send_message("test message", channel = "My Channel")
}
}
