% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello.R
\name{trello_upload}
\alias{trello_upload}
\alias{trello_credentials}
\alias{trello_getboards}
\alias{trello_getlists}
\alias{trello_getmembers}
\alias{trello_getcards}
\alias{trello_getlabels}
\alias{trello_getcustomfields}
\alias{trello_getchecklists}
\alias{trello_getcomments}
\alias{trello_searchcard}
\alias{trello_openboard}
\alias{trello_opencard}
\alias{trello_get_card_property}
\alias{trello_get_card_id}
\alias{trello_setcomment}
\alias{trello_setdeadline}
\alias{trello_addtask}
\alias{trello_settask_state}
\alias{trello_movecard}
\title{Work with Trello Cards}
\usage{
trello_upload(
  title,
  desc = "",
  requested_by = "",
  project_path = "",
  member = Sys.getenv("R_USERNAME"),
  prio = read_secret("trello.default.prio"),
  duedate = "",
  checklist = "",
  checklist_name = read_secret("trello.default.checklist"),
  comments = "",
  attachments = "",
  list = read_secret("trello.default.list"),
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token"),
  ...
)

trello_credentials(x = c("member", "key", "token"), item = 1)

trello_getboards(
  username = trello_credentials("member"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getlists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getmembers(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcards(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getlabels(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcustomfields(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getchecklists(
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_getcomments(
  card_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_searchcard(
  x,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_openboard(board = read_secret("trello.default.board"))

trello_opencard()

trello_get_card_property(
  card_number,
  property,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_get_card_id(
  card_number,
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_setcomment(
  card_id,
  comment,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_setdeadline(
  card_id,
  duedate,
  duecomplete = FALSE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_addtask(
  card_id,
  new_items_vector = NULL,
  checklist_name = "Taken",
  board = read_secret("trello.default.board"),
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_settask_state(
  card_id,
  checkitem_id,
  new_value = TRUE,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)

trello_movecard(
  card_id,
  list_id,
  key = trello_credentials("key"),
  token = trello_credentials("token")
)
}
\arguments{
\item{title}{card title}

\item{desc}{card description}

\item{requested_by}{name of requester}

\item{project_path}{path to project to which this card belongs to}

\item{member}{Trello member, full name}

\item{prio}{priority, must be label name}

\item{duedate}{deadline}

\item{checklist}{items for checklist, can be character vector}

\item{checklist_name}{name for the checklist}

\item{comments}{comment to put on card, can be character vector}

\item{attachments}{attachments to put on card}

\item{list}{Trello list}

\item{board}{Trello board}

\item{key}{Trello API key}

\item{token}{Trello API token}

\item{...}{arguments passed on to methods}
}
\description{
This functions all call the \href{https://api.trello.com}{Trello api}.
}
